// arg-web-frontend/src/types/accounts-payable.types.ts
export interface Update1099FileData {
  key: string;
  recordType: string;
  ctl: string;
  tin: string;
  firstPayeeName: string;
}
// ...
  export interface ApPeriodEndReportItem {
  recordType?: string;
  ctl?: string;
  tin?: string;
  firstPayeeName?: string;
}