// arg-web-frontend/src/api/api-schema/api.ts
export interface ApPeriodEndDto {
  /**
   * Tin Number
   * @example "10"
   */
  tin: string;
  /**
   * Ctl
   * @example "10"
   */
  ctl: string;
}